Revert "RATIS-2099. Cache TermIndex"
This reverts commit 3943a2b9aa6cf1a2b23ca7e7cdf804455a8f864e.
